Move InteractionHook to interactions package
